Message detail object refactoring ClientDemandMessageDetail - new attributes
git-svn-id: https://svn.webgres.cz/svn/poptavka@840 61682b0c-635e-44ba-a85b-d27e7dc34134

+++ poptavka-core/src/main/java/cz/poptavka/sample/dao/message/MessageDao.java
@@ -92,7 +92,7 @@ public interface MessageDao extends GenericDao<Message> {
      * of ALL the messages spanning from the demand message (including the
      * demand message itself)
      */
-    Map<Message, Long> getListOfClientDemandMessagesAll(User user);
+    Map<Message, Integer> getListOfClientDemandMessagesAll(User user);
 
     /**
      * Gets all the demand messages of the given user along with the number of
@@ -104,7 +104,7 @@ public interface MessageDao extends GenericDao<Message> {
      * of UNREAD messages spanning from the demand message (including the
      * demand message itself)
      */
-    Map<Message, Long> getListOfClientDemandMessagesUnread(User user);
+    Map<Message, Integer> getListOfClientDemandMessagesUnread(User user);
 
     /**
      * Gets all the descendants (not just the children) of every item
